requirements.txt
requests~=2.32.4
Flask~=3.1.1
aiogram~=3.21.0
python-dotenv~=1.1.1